@@ -8,8 +8,9 @@
 # serve to show the default.
 
 import os
-import sys
 import shutil
+import sys
+from contextlib import suppress
 
 # -- Path setup --------------------------------------------------------------
 
@@ -35,10 +36,9 @@ except ImportError:
 
 output_dir = os.path.join(__location__, "api")
 module_dir = os.path.join(__location__, "../src/strawberry_sqlalchemy_mapper")
-try:
+
+with suppress(FileNotFoundError):
     shutil.rmtree(output_dir)
-except FileNotFoundError:
-    pass
 
 try:
     import sphinx
@@ -52,7 +52,7 @@ try:
 
     apidoc.main(args)
 except Exception as e:
-    print("Running `sphinx-apidoc` failed!\n{}".format(e))
+    print(f"Running `sphinx-apidoc` failed!\n{e}")  # noqa: T201
 
 # -- General configuration ---------------------------------------------------
 
@@ -96,7 +96,7 @@ copyright = "2022, Tim Dumol"
 #
 # version: The short X.Y version.
 # release: The full version, including alpha/beta/rc tags.
-# If you don’t need the separation provided between version and release,
+# If you don't need the separation provided between version and release,
 # just set them both to the same value.
 try:
     from strawberry_sqlalchemy_mapper import __version__ as version
@@ -158,10 +158,7 @@ html_theme = "alabaster"
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-html_theme_options = {
-    "sidebar_width": "300px",
-    "page_width": "1200px"
-}
+html_theme_options = {"sidebar_width": "300px", "page_width": "1200px"}
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []
@@ -246,7 +243,13 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-    ("index", "user_guide.tex", "strawberry-sqlalchemy-mapper Documentation", "Tim Dumol", "manual")
+    (
+        "index",
+        "user_guide.tex",
+        "strawberry-sqlalchemy-mapper Documentation",
+        "Tim Dumol",
+        "manual",
+    )
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -283,4 +286,4 @@ intersphinx_mapping = {
     "pyscaffold": ("https://pyscaffold.org/en/stable", None),
 }
 
-print(f"loading configurations for {project} {version} ...", file=sys.stderr)
+print(f"loading configurations for {project} {version} ...", file=sys.stderr)  # noqa: T201
