@@ -4,7 +4,7 @@ from nox_poetry import Session, session
 nox.options.reuse_existing_virtualenvs = True
 nox.options.error_on_external_run = True
 
-PYTHON_VERSIONS = ["3.12", "3.11", "3.10", "3.9", "3.8"]
+PYTHON_VERSIONS = ["3.13", "3.12", "3.11", "3.10", "3.9", "3.8"]
 
 
 COMMON_PYTEST_OPTIONS = [
@@ -16,6 +16,10 @@ COMMON_PYTEST_OPTIONS = [
     "--showlocals",
     "-vv",
 ]
+
+
+def poetry_install_run_always(session: Session) -> None:
+    session.run_always("poetry", "install", external=True)
 
 
 @session(python=PYTHON_VERSIONS, name="SQLAlchemy 2.0 Tests", tags=["tests"])
@@ -31,7 +35,7 @@ def tests_sqlalchemy_latest(session: Session) -> None:
         "wheel",
         external=True,
     )
-    session.run_always("poetry", "install", external=True)
+    poetry_install_run_always(session)
 
     session.run(
         "pytest",
@@ -53,7 +57,7 @@ def tests_sqlalchemy_1_4(session: Session) -> None:
         "wheel",
         external=True,
     )
-    session.run_always("poetry", "install", external=True)
+    poetry_install_run_always(session)
     session._session.install("sqlalchemy~=1.4")
 
     session.run(
@@ -64,6 +68,40 @@ def tests_sqlalchemy_1_4(session: Session) -> None:
 
 @session(name="Mypy", tags=["lint"])
 def mypy(session: Session) -> None:
-    session.run_always("poetry", "install", external=True)
+    poetry_install_run_always(session)
 
-    session.run("mypy", "--config-file", "mypy.ini")
+    session.run(
+        "mypy",
+        "--install-types",
+        "--non-interactive",
+        "--cache-dir=.mypy_cache/",
+        "--config-file",
+        "mypy.ini",
+    )
+
+
+@session(name="Ruff Lint", tags=["lint"])
+def ruff_lint(session: Session) -> None:
+    poetry_install_run_always(session)
+
+    session.run(
+        "ruff",
+        "check",
+        "--no-fix",
+        ".",
+        silent=False,
+    )
+
+
+@session(name="Ruff Format", tags=["format"])
+def ruff_format(session: Session) -> None:
+    poetry_install_run_always(session)
+
+    session.run(
+        "ruff",
+        "format",
+        "--check",
+        "--diff",
+        ".",
+        silent=False,
+    )
