@@ -13,11 +13,6 @@ from strawberry_sqlalchemy_mapper import StrawberrySQLAlchemyMapper
 
 
 @pytest.fixture
-def mapper():
-    return StrawberrySQLAlchemyMapper()
-
-
-@pytest.fixture
 def polymorphic_employee(base):
     class Employee(base):
         __tablename__ = "employee"
@@ -80,16 +75,13 @@ def polymorphic_employee_table(base):
 
 def test_mapper_default_model_to_type_name(employee_table):
     Employee = employee_table
-    assert (
-        StrawberrySQLAlchemyMapper._default_model_to_type_name(Employee) == "Employee"
-    )
+    assert StrawberrySQLAlchemyMapper._default_model_to_type_name(Employee) == "Employee"
 
 
 def test_default_model_to_interface_name(employee_table):
     Employee = employee_table
     assert (
-        StrawberrySQLAlchemyMapper._default_model_to_interface_name(Employee)
-        == "EmployeeInterface"
+        StrawberrySQLAlchemyMapper._default_model_to_interface_name(Employee) == "EmployeeInterface"
     )
 
 
@@ -142,18 +134,16 @@ def test_convert_all_columns_to_strawberry_type(mapper):
         strawberry_type,
     ) in mapper.sqlalchemy_type_to_strawberry_type_map.items():
         assert (
-            mapper._convert_column_to_strawberry_type(
-                Column(sqlalchemy_type, nullable=False)
-            )
+            mapper._convert_column_to_strawberry_type(Column(sqlalchemy_type, nullable=False))
             == strawberry_type
         )
 
 
 def test_convert_column_to_strawberry_type(mapper):
     int_column = Column(Integer, nullable=False)
-    assert mapper._convert_column_to_strawberry_type(int_column) == int
+    assert mapper._convert_column_to_strawberry_type(int_column) is int
     string_column = Column(String, nullable=False)
-    assert mapper._convert_column_to_strawberry_type(string_column) == str
+    assert mapper._convert_column_to_strawberry_type(string_column) is str
 
 
 def test_convert_json_column_to_strawberry_type(mapper):
@@ -180,9 +170,7 @@ def test_convert_enum_column_to_strawberry_type(mapper):
     assert mapper._convert_column_to_strawberry_type(column) == SampleEnum
 
 
-def test_convert_relationship_to_strawberry_type(
-    employee_and_department_tables, mapper
-):
+def test_convert_relationship_to_strawberry_type(employee_and_department_tables, mapper):
     _, Department = employee_and_department_tables
     employees_property = Department.employees.property
     assert (
@@ -231,14 +219,12 @@ def test_type_simple(employee_table, mapper):
     assert len(mapped_employee_type.__strawberry_definition__.fields) == 2
     employee_type_fields = mapped_employee_type.__strawberry_definition__.fields
     name = next(iter(filter(lambda f: f.name == "name", employee_type_fields)))
-    assert name.type == str
+    assert name.type is str
     id = next(iter(filter(lambda f: f.name == "id", employee_type_fields)))
-    assert id.type == int
+    assert id.type is int
 
 
-def test_interface_and_type_polymorphic(
-    mapper, polymorphic_employee, polymorphic_lawyer
-):
+def test_interface_and_type_polymorphic(mapper, polymorphic_employee, polymorphic_lawyer):
     @mapper.interface(polymorphic_employee)
     class EmployeeInterface:
         pass
@@ -284,9 +270,7 @@ def test_use_list(employee_and_department_tables, mapper):
     assert len(mapped_department_type.__strawberry_definition__.fields) == 3
     department_type_fields = mapped_department_type.__strawberry_definition__.fields
 
-    name = next(
-        (field for field in department_type_fields if field.name == "employees"), None
-    )
+    name = next((field for field in department_type_fields if field.name == "employees"), None)
     assert name is not None
     assert isinstance(name.type, StrawberryOptional) is False
     assert isinstance(name.type, StrawberryList) is True
@@ -307,9 +291,9 @@ def test_type_relationships(employee_and_department_tables, mapper):
     assert len(mapped_employee_type.__strawberry_definition__.fields) == 4
     employee_type_fields = mapped_employee_type.__strawberry_definition__.fields
     name = next(iter(filter(lambda f: f.name == "department_id", employee_type_fields)))
-    assert type(name.type) == StrawberryOptional
+    assert type(name.type) is StrawberryOptional
     id = next(iter(filter(lambda f: f.name == "department", employee_type_fields)))
-    assert type(id.type) == StrawberryOptional
+    assert type(id.type) is StrawberryOptional
 
 
 def test_relationships_schema(employee_and_department_tables, mapper):
@@ -335,7 +319,7 @@ def test_relationships_schema(employee_and_department_tables, mapper):
     type Department {
       id: Int!
       name: String!
-      employees: EmployeeConnection!
+      employees(first: Int = null, after: String = null, last: Int = null, before: String = null): EmployeeConnection!
     }
 
     type Employee {
@@ -377,5 +361,83 @@ def test_relationships_schema(employee_and_department_tables, mapper):
     type Query {
       departments: Department!
     }
-    '''
+    '''  # noqa: E501 - long lines needed for exact string matches
     assert str(schema) == textwrap.dedent(expected).strip()
+
+
+@pytest.mark.parametrize(
+    "directives",
+    [
+        (["@deprecated(reason: 'Use newEmployee instead')"]),
+        (
+            [
+                "@deprecated(reason: 'Use newEmployee instead')",
+                "@customDirective(value: 'example')",
+            ]
+        ),
+    ],
+)
+def test_type_with_directives(mapper, employee_table, directives):
+    Employee = employee_table
+
+    @mapper.type(Employee, directives=directives)
+    class Employee:
+        pass
+
+    mapper.finalize()
+    additional_types = list(mapper.mapped_types.values())
+    assert len(additional_types) == 1
+    mapped_employee_type = additional_types[0]
+    assert mapped_employee_type.__name__ == "Employee"
+    assert len(mapped_employee_type.__strawberry_definition__.fields) == 2
+    assert mapped_employee_type.__strawberry_definition__.directives == directives
+
+
+@pytest.mark.parametrize(
+    "directives",
+    [
+        (["@deprecated(reason: 'Use newEmployee instead')"]),
+        (
+            [
+                "@deprecated(reason: 'Use newEmployee instead')",
+                "@customDirective(value: 'example')",
+            ]
+        ),
+    ],
+)
+def test_type_with_directives_and_federation(mapper, employee_table, directives):
+    Employee = employee_table
+
+    @mapper.type(Employee, directives=directives, use_federation=True)
+    class Employee:
+        pass
+
+    mapper.finalize()
+    additional_types = list(mapper.mapped_types.values())
+    assert len(additional_types) == 1
+    mapped_employee_type = additional_types[0]
+    assert mapped_employee_type.__name__ == "Employee"
+    assert len(mapped_employee_type.__strawberry_definition__.fields) == 2
+    assert mapped_employee_type.__strawberry_definition__.directives == directives
+
+
+@pytest.mark.parametrize(
+    ("use_federation_value", "expected_directives"),
+    [(True, []), (False, ())],
+)
+def test_type_with_default_directives(
+    mapper, employee_table, use_federation_value, expected_directives
+):
+    Employee = employee_table
+
+    @mapper.type(Employee, use_federation=use_federation_value)
+    class Employee:
+        pass
+
+    mapper.finalize()
+    additional_types = list(mapper.mapped_types.values())
+    assert len(additional_types) == 1
+    mapped_employee_type = additional_types[0]
+    assert mapped_employee_type.__name__ == "Employee"
+    assert len(mapped_employee_type.__strawberry_definition__.fields) == 2
+    assert mapped_employee_type.__strawberry_definition__.directives == expected_directives
